+++ config/webpack/webpack.config.js
@@ -57,7 +57,7 @@ const webpackConfig = (options, env, argv) => {
             {
               loader: 'postcss-loader',
               options: {
-                // Add unique ident to prevent the loader from searching for a postcss.config file. Additionally see: https://github.com/postcss/postcss-loader#plugins
+                // Add unique ident to prevent the loader from searching for a postcss.config file. See: https://github.com/postcss/postcss-loader#plugins
                 ident: 'postcss',
                 sourceMap: true,
                 plugins() {