+++ tests/jest/aggregate-themes.test.js
@@ -2,7 +2,6 @@ const ThemeAggregator = require('../../scripts/aggregate-themes/theme-aggregator
 
 global.console = { log: jest.fn(), warn: jest.fn() };
 
-
 describe('Theme Aggregator', () => {
   describe('aggregate', () => {
     it('returns null if there is no terra-theme-config.js', () => {
@@ -94,7 +93,7 @@ describe('Theme Aggregator', () => {
       const options = { theme: 'terra-mock-dark-theme' };
 
       ThemeAggregator.validate(options);
-
+      // eslint-disable-next-line no-console
       expect(console.warn).toHaveBeenCalledTimes(0);
     });
 
@@ -102,7 +101,7 @@ describe('Theme Aggregator', () => {
       const options = { scoped: ['terra-mock-dark-theme'] };
 
       ThemeAggregator.validate(options);
-
+      // eslint-disable-next-line no-console
       expect(console.warn).toHaveBeenCalledTimes(0);
     });
 
@@ -110,7 +109,7 @@ describe('Theme Aggregator', () => {
       const options = {};
 
       ThemeAggregator.validate(options);
-
+      // eslint-disable-next-line no-console
       expect(console.warn).toHaveBeenCalled();
     });
   });