+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Added 
-* Added default testName in themeCombinationOfCustomProperties
+* Added default testName `themed` for `themeCombinationOfCustomProperties` helper.
 
 4.27.0 - (April 16, 2019)
 ----------