+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Changelog
 Unreleased
 ----------
 
+### Added
+* Added default testName `themed` for `themeCombinationOfCustomProperties` helper.
+* Add `terra-aggregate-translations` module
+* Guard against empty screenshot array being passed to `getComparisonResults`.
+
+### Remove
+* Aggregate-translation script, tests, and bin executable
+
 4.27.0 - (April 16, 2019)
 ----------
 ### Added